Feat: Add a memory pool to store malloced memory
In functions that request for memory with malloc, but do not themselves
return that memory (so it can be directly freed in main--or some other
function), a memory pool (a.k.a mem pool) will be utilized to store
those memory locations so that it can later be freed properly.

The mem pool has to be passed into such a function as a parameter for
the strategy to work.