Refactor: Replace `concat_all` with `concat_three`
`concat_all` was too general a function for no good reason (and it did
not play well with valgrind as it repeatedly reported uninitialized
value(s)).

`concat_three` replaces `concat_all` while still retaining the
flexibility of concatenating up to three strings.

Even though this puts a hard limit on how many strings can be concatenated,
it eliminates the valgrind messages and is much simpler to understand.