Feat: Add a "mem pool" to free tricky memory
"Tricky memory" refers to malloced memory that cannot easily be freed
inside the function that directly or inderectly requested that memory.

A function that returns an integer, but requests memory that should
stick around for the lifetime of the program (and therefore cannot free
that memory within itself), is one such scenario.

A memory pool, that is passed as an argument to such a function solves
this dilema by keeping track of any malloced memory that can later be
freed by freeing the memory pool later.